refactor: move normalizeQuestionsData to separate utility file
Move the normalization function to src/utils/questions.ts to avoid
test import issues with complex dependencies in api.ts.
